pyproject.toml
[tool.poetry]
name = "fastapi-audit-log"
version = "0.2.0"
description = "Production-grade, reusable audit logging for FastAPI applications."
authors = ["Ahmed Elgamal <ahmed@example.com>"]
readme = "README.md"
packages = [{include = "fastapi_audit_log"}]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = ">=0.100.0"
pydantic = "^2.0"
pydantic-settings = ">=2.0"
asyncpg = { version = ">=0.29", optional = true }
aiosqlite = { version = "*", optional = true }
aiomysql = { version = "*", optional = true }
sqlalchemy = { version = ">=2.0.0", extras = ["asyncio"], optional = true }
tortoise-orm = { version = ">=0.20", optional = true }
sqlmodel = { version = ">=0.0.14", optional = true }
beanie = { version = ">=1.20", optional = true }
motor = { version = ">=3.0", optional = true }

[tool.poetry.extras]
sqlalchemy = ["sqlalchemy", "asyncpg", "aiomysql", "aiosqlite"]
tortoise = ["tortoise-orm", "asyncpg", "aiomysql", "aiosqlite"]
sqlmodel = ["sqlmodel", "asyncpg", "aiomysql", "aiosqlite"]
mongodb = ["beanie", "motor"]
asyncpg = ["asyncpg"]
all = ["sqlalchemy", "tortoise-orm", "sqlmodel", "beanie", "motor", "asyncpg", "aiomysql", "aiosqlite"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.27.0"
pytest-cov = "^4.1.0"
ruff = "^0.3.0"
mypy = "^1.9.0"
uvicorn = "^0.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "DTZ", "T20", "RET", "SIM", "ARG", "PTH", "ERA"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"examples/*" = ["T201", "ARG001"]
"tests/*" = ["ARG001"]

[tool.mypy]
strict = true
python_version = "3.11"
plugins = ["pydantic.mypy"]
exclude = ["tests", "examples"]
